+++ CHANGELOG.md
@@ -33,6 +33,7 @@ BUG FIXES:
 - `pg` backend doesn't fail on workspace creation for paralel runs, when the database is shared across multiple projects. ([#2411](https://github.com/opentofu/opentofu/pull/2411))
 - Generating an OpenTofu configuration from an `import` block that is referencing a resource with nested attributes now works correctly, instead of giving an error that the nested computed attribute is required. ([#2372](https://github.com/opentofu/opentofu/issues/2372)) 
 - `base64gunzip` now doesn't expose sensitive values if it fails during the base64 decoding. ([#2503](https://github.com/opentofu/opentofu/pull/2503))
+- Backend `azurerm` correctly handles `metadata_host` value for custom cloud environments ([#2512](https://github.com/opentofu/opentofu/pull/2512))
 
 ## Previous Releases